// src/app/models/funcionario.model.ts
export interface Funcionario {
  id: number;
  tipoDocumento: string;
  apellidoFuncionario: string;
  cargo: string;
  celular: string;
  direccion: string;
  email: string;
  estado: string;
  estadoCivil: string;
  fechaNacimiento: Date;
  genero: string;
  identificacion: string;
  nombreFuncionario: string;
}